+++ src/store/StepProgress/StepProgressContext.tsx
@@ -1,3 +1,4 @@
+import { Client } from '@/app/FormView/services/addClient.service'
 import { StepProgress } from '@/types/StepProgress'
 import { createContext } from 'react'
 
@@ -8,6 +9,8 @@ interface IStepProgressContext {
   setTotalSubsteps: (totalSubsteps: number) => void
   setCurrentSubstep: (currentSubstep: number) => void
   setStepProgress: React.Dispatch<React.SetStateAction<StepProgress>>
+  clientData: Client,
+  setClientData: (data:Client) => void
 }
 
 export const StepProgressContext = createContext<IStepProgressContext>({
@@ -22,4 +25,15 @@ export const StepProgressContext = createContext<IStepProgressContext>({
   setTotalSubsteps: () => {},
   setCurrentSubstep: () => {},
   setStepProgress: () => {},
+  clientData:{
+    email: '',
+    dni: '',
+    telf: '',
+    nombres: '',
+    primerApl: '',
+    segundoApl: '',
+    countryState: '',
+    city: ''
+  },
+  setClientData: () => {}
 })
